refactor: 비즈니스 로직을 훅으로 분리
